Refactor corrections ui into components
