+++ frontend/src/screens/App/screens/Corrections/components/CorrectAddress.less
@@ -1,0 +1,3 @@
+.addressInput {
+  width: 100%;
+}
