+++ frontend/src/screens/App/screens/Corrections/components/CorrectGeocode.less
@@ -1,0 +1,80 @@
+@import '../../../../../shared/styles/fonts.less';
+@import '../../../../../shared/styles/colors.less';
+@import '../../../../../shared/styles/viewport.less';
+@import '../../../../../shared/styles/utils.less';
+
+.coordinateInputRow {
+  display: grid;
+  grid-template-columns: 1fr auto 1fr 1fr;
+  align-items: flex-end;
+
+  .openMapButton {
+    margin: 0 0.5rem;
+    padding: 0 0.5rem;
+    align-self: stretch;
+  }
+}
+
+.alternates {
+  text-align: initial;
+  border: none;
+  margin: 1rem 0;
+  padding: 0;
+
+  legend {
+    font-weight: bold;
+  }
+}
+
+.alternatesCheckboxes {
+  margin-top: 1rem;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, 11px * 10);
+  grid-auto-rows: 11px * 7.5;
+  grid-gap: 11px;
+
+  & > input[type='checkbox'] {
+    .visuallyhidden();
+  }
+
+  & > input[type='checkbox'] + label {
+    position: relative;
+    box-sizing: border-box;
+    border: 2px solid @font-color-dark;
+    border-radius: 4px;
+    cursor: pointer;
+
+    &::after {
+      content: '✗';
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 11px * 2;
+      height: 11px * 2;
+      text-align: center;
+      color: white;
+      background-color: @font-color-dark;
+    }
+
+    .thumbnail {
+      filter: brightness(0.5);
+    }
+  }
+
+  & > input[type='checkbox']:checked + label {
+    &::after {
+      content: '✓';
+    }
+
+    .thumbnail {
+      filter: brightness(1);
+    }
+  }
+
+  .thumbnail {
+    display: block;
+    object-fit: cover;
+    height: 100%;
+    width: 100%;
+  }
+}
