+++ frontend/src/screens/App/screens/Corrections/index.tsx
@@ -1,10 +1,5 @@
 import React from 'react';
 
-import classNames from 'classnames';
-import isEmpty from 'lodash/isEmpty';
-import map from 'lodash/map';
-import uniqueId from 'lodash/uniqueId';
-
 import useLoginStore from 'shared/stores/LoginStore';
 import useCorrectionsStore, {
   useCorrectionsStoreComputeds,
@@ -14,41 +9,17 @@ import Button from 'shared/components/Button';
 import FieldSet from 'shared/components/FieldSet';
 import LoginForm from 'shared/components/LoginForm';
 import FourtiesModal from 'shared/components/Modal';
-import TextInput from 'shared/components/TextInput';
+import CorrectAddress from './components/CorrectAddress';
+import CorrectGeocode from './components/CorrectGeocode';
 import PhotoMetadata from './components/PhotoMetadata';
 
-import { PHOTO_BASE } from 'shared/utils/apiConstants';
-import CoordinateInput from './components/CoordinateInput';
-import LocationPickerModal from './components/LocationPickerModal';
 import stylesheet from './Corrections.less';
 
 const CorrectionsDialogContent = (): JSX.Element | null => {
-  const {
-    photo,
-    alternatesSelections,
-    toggleAlternateSelection,
-    openMap,
-    correctedLng,
-    correctedLat,
-    setCorrectedLngLat,
-    correctedAddress,
-    setCorrectedAddress,
-    submit,
-  } = useCorrectionsStore();
-  const {
-    previousLng: defaultLng,
-    previousLat: defaultLat,
-    previousAddress: defaultAddress,
-    canSubmit,
-  } = useCorrectionsStoreComputeds();
+  const { photo, submit } = useCorrectionsStore();
+  const { canSubmit } = useCorrectionsStoreComputeds();
 
   const { isLoginValidated } = useLoginStore();
-
-  const handleAlternateSelectionChange = (identifier: string): void => {
-    toggleAlternateSelection(identifier);
-  };
-
-  const idPrefix = uniqueId('corrections-');
 
   return (
     <div className={stylesheet.modalContent}>
@@ -86,116 +57,8 @@ const CorrectionsDialogContent = (): JSX.Element | null => {
           <LoginForm />
         </FieldSet>
         <div className={stylesheet.correctionsGroup}>
-          <FieldSet
-            className={stylesheet.fieldset}
-            disabled={!isLoginValidated}
-          >
-            <FieldSet.Legend>Correct map placement</FieldSet.Legend>
-
-            <p>Fix where the dot for this photo appears on the map.</p>
-
-            <div className={stylesheet.coordinateInputRow}>
-              <CoordinateInput
-                name="lat"
-                label="Latitude"
-                placeholder={defaultLat}
-                value={correctedLat}
-                onValueChange={(newValue) =>
-                  setCorrectedLngLat(correctedLng ?? null, newValue)
-                }
-                rangeMin={-90}
-                rangeMax={90}
-              />
-              <span style={{ verticalAlign: 'bottom' }}>, </span>
-              <CoordinateInput
-                name="lng"
-                label="Longitude"
-                placeholder={defaultLng}
-                value={correctedLng}
-                onValueChange={(newValue) =>
-                  setCorrectedLngLat(newValue ?? null, correctedLat)
-                }
-                rangeMin={-180}
-                rangeMax={180}
-              />
-              <Button
-                buttonStyle="primary"
-                buttonTheme="modal"
-                className={stylesheet.openMapButton}
-                onClick={openMap}
-              >
-                Select on map
-              </Button>
-              <LocationPickerModal />
-            </div>
-
-            {!isEmpty(alternatesSelections) ? (
-              <fieldset className={stylesheet.alternates}>
-                <legend>
-                  Also move these other photos at the same location?
-                </legend>
-                <p>
-                  If they are not moved together they will be dissociated. 80s
-                  photos left without a 40s photo at the same location are
-                  hidden from the site.
-                </p>
-                <div>
-                  <div className={stylesheet.alternatesCheckboxes}>
-                    {map(alternatesSelections, (selected, identifier) => (
-                      <React.Fragment key={identifier}>
-                        <input
-                          id={`${idPrefix}-${identifier}`}
-                          type="checkbox"
-                          name={identifier}
-                          checked={selected}
-                          onChange={() => {
-                            handleAlternateSelectionChange(identifier);
-                          }}
-                          className={stylesheet.alternateCheckbox}
-                        />
-                        <label htmlFor={`${idPrefix}-${identifier}`}>
-                          <img
-                            alt={identifier}
-                            src={`${PHOTO_BASE}/420-jpg/${identifier}.jpg`}
-                            className={stylesheet.thumbnail}
-                            title={identifier}
-                            draggable={false}
-                          />
-                        </label>
-                      </React.Fragment>
-                    ))}
-                  </div>
-                </div>
-              </fieldset>
-            ) : (
-              <p>
-                <i>No other photos at this location.</i>
-              </p>
-            )}
-          </FieldSet>
-
-          <FieldSet
-            className={stylesheet.fieldset}
-            disabled={!isLoginValidated}
-          >
-            <FieldSet.Legend>Correct address</FieldSet.Legend>
-
-            <p>Fix the address sometimes displayed with this photo.</p>
-
-            <div className={stylesheet.addressInputRow}>
-              <TextInput
-                type="text"
-                name="address"
-                aria-label="Address"
-                className={classNames(stylesheet.addressInput)}
-                placeholder={defaultAddress}
-                value={correctedAddress}
-                onChange={(event) =>
-                  setCorrectedAddress(event.target.value || null)
-                }
-              />
-            </div>
-          </FieldSet>
+          <CorrectGeocode />
+          <CorrectAddress />
         </div>
 
         <p>
