+++ frontend/src/screens/App/screens/CreditPurchaseModal/index.tsx
@@ -126,4 +126,6 @@ export function openCreditPurchaseModal(): void {
   useCreditPurchaseModalStore.getState().open();
 }
 
+openCreditPurchaseModal();
+
 export { default as CreditPurchaseSuccessMessage } from './SuccessMessage';
